change uint16 to uint32

+++ INCDIR/Hboard/boarddrv.h
@@ -62,10 +62,10 @@ INT16 SciaRx_Ready(void);
 INT16 SciaTx_Ready(void);
 void UartPrintf(const char *fmt, ...);
 UINT8 UartGetc(void);
-UINT8  UartCharGetTimeout(UINT16 timeout, int *result);
+UINT8  UartCharGetTimeout(UINT32 timeout, int *result);
 void UartPutbuffer(UINT8 *s);
 
-UINT8 UartCharGetTimeout_B(UINT16 timeout, int *result);
+UINT8 UartCharGetTimeout_B(UINT32 timeout, int *result);
 int UartWrite_B(UINT8 *buffer, UINT16 length);
 int UartRead_B(UINT8* buffer, UINT16 length);
